Adding support for Redis backend
Why wasn't redis support there to begin with?

This is just the first pass, so there is still more stuff to clean up:
  * console statements
  * config for redis

The basic idea behind this is that we open a new connection to redis to
subscribe on the keyspace for this database.
And, everytime we get a 'set' event, we assume that the set was done as
a way of completing a Celery task.
So, since the listener/subscription for these events are on the client,
then the client will fire off another "taskFinished" event, which the
result will listen on to finally set a task to "ready".

Now that i'm typing this out, it seems kinda messy...